# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  auth.ErrorResponse:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        description: Details is present only for VALIDATION_ERROR responses.
      message:
        example: Human-readable error description.
        type: string
    type: object
...
  notifications.NotificationListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/notifications.NotificationRead'
        type: array
      total:
        type: integer
      unread_count:
        type: integer
    type: object
  notifications.NotificationRead:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      message:
        type: string
      payload:
        items:
          type: integer
        type: array
      read_at:
        type: string
      title:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  notifications.SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
  notifications.UnreadCountResponse:
    properties:
      unread_count:
        type: integer
    type: object
host: localhost:8000
info:
  contact:
    email: dev@flickey.com
    name: Flickey Dev Team
  description: Property-rental platform backend. Provides OTP-based authentication,
    multi-step listing wizard, media upload workflow, amenity catalogue, and an admin
    moderation panel.
  license:
    name: MIT
  termsOfService: http://flickey.com/terms/
  title: Flickey Go API
  version: "1.0"
paths:
...
  /notifications:
    get:
      description: Returns a paginated list of notifications for the authenticated
        user.
      parameters:
      - description: Limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by unread only
        in: query
        name: unread_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notifications.NotificationListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user notifications
      tags:
      - Notifications
  /notifications/{id}/read:
    patch:
      description: Marks a specific notification as read.
      parameters:
      - description: Notification UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notifications.SuccessResponse'
        "400":
          description: INVALID_ID
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - Notifications
  /notifications/read-all:
    post:
      description: Marks all unread notifications for the user as read.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notifications.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - Notifications
  /notifications/stream:
    get:
      description: Establishes a persistent SSE stream to deliver real-time notifications
        to the client. Token can be passed in Authorization header or as ?token= query
        param.
      parameters:
      - description: JWT access token (for EventSource)
        in: query
        name: token
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: 'event: notification\ndata: {...}\n\n'
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Stream real-time notifications (SSE)
      tags:
      - Notifications
  /notifications/unread-count:
    get:
      description: Returns total count of unread notifications for the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notifications.UnreadCountResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get unread notifications count
      tags:
      - Notifications
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: **Bearer {token}**'
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: HttpOnly refresh token cookie set by /auth/verify-otp and rotated
      by /auth/refresh.
    in: cookie
    name: refresh_token
    type: apiKey
swagger: "2.0"